refactor: UI 관련 클래스 패키지 이동 및 네이밍 변경
