+++ app/src/main/java/woowacourse/payments/AddCardActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import woowacourse.payments.domain.model.toUiModel
+import woowacourse.payments.ui.mapper.toUiModel
 import woowacourse.payments.ui.screen.AddCardScreen
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
